+++ inlang/source-code/paraglide/paraglide-js/src/compiler/compileMessage.ts
@@ -29,10 +29,10 @@ export const compileMessage = (message: Message): string => {
 
 	return `
 /**
- * This message has been compiled by [inlang paraglide](https://inlang.com/marketplace/library.inlang.paraglideJs).
+ * This message has been compiled by [inlang paraglide](https://inlang.com/m/gerre34r/library-inlang-paraglideJs).
  *
- * - Don't edit the message manually. Use the [inlang ide extension](https://inlang.com/marketplace/app.inlang.ideExtension)
- *   or the [web editor](https://inlang.com/marketplace/app.inlang.editor) to edit the message.
+ * - Don't edit the message's code. Use the [inlang ide extension](https://inlang.com/m/r7kp499g/app-inlang-ideExtension),
+ *   the [web editor](https://inlang.com/m/tdozzpar/app-inlang-editor) instead, or edit the translation files manually.
  * 
  * - The params are NonNullable<unknown> because inlang can't know the value type of a param (yet).
  * 
@@ -40,6 +40,7 @@ export const compileMessage = (message: Message): string => {
  * @returns {string}
  */
   export const ${message.id} = (${Object.keys(params).length > 0 ? "params" : ""}) => {
+		/** @type {Record<string, string>} */
     const variants = {
       ${Object.entries(contents)
 				.map(