+++ inlang/source-code/website/src/interface/marketplace/categoryHeaders/application/svelte.tsx
@@ -1,6 +1,6 @@
 import { IconSvelte } from "#src/interface/custom-icons/subcategoryIcon.jsx"
 import { Button } from "#src/pages/index/components/Button.jsx"
-import * as m from "@inlang/paraglide-js/website/messages"
+import * as m from "#src/paraglide/messages.js"
 
 const SvelteHeader = () => {
 	return (