+++ inlang/source-code/website/vite.config.ts
@@ -32,7 +32,7 @@ export default defineConfig({
 		// markdownHotModuleReload(),
 		watch({
 			pattern: "static/messages.json",
-			command: "paraglide-js compile --namespace website --project ../../../project.inlang.json",
+			command: "paraglide-js compile --project ../../../project.inlang.json",
 		}),
 	],
 	resolve: {